static files

+++ src/com/esimorp/shs/entity/body/HttpBody.java
@@ -2,7 +2,7 @@ package com.esimorp.shs.entity.body;
 
 
 public abstract class HttpBody {
-    public abstract int getContentLength();
+    public abstract long getContentLength();
 
     public abstract String getContentType();
